requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pillow>=10.0.0
jinja2>=3.1.0
python-slugify>=8.0.0
psutil>=5.9.0

